Ok, this demonstrates how I was messing up pointer stuff
So inside the function we allocated a Solver struct with a struct
literal. Then we started mutating it and its contained Cell structs.
Those were Cell values in arrays, with *board pointers to the local
Solver. Then we returned that solver.

Now I realize before the function was even called, a Solver struct was
also allocated in the calling function, when the variable was declared.
Then the function returned a Solver struct, which was copied into the
variable's location in memory. Because `cells` was an array, not a
slice, each element was copied. Becase each Cell was a struct, not a
pointer, each Cell was copied. So, the inner Solver's Cells were all
copied into the outer Solver.

But, those Cells all have pointers to the original, inner Solver! That
was the source of the weirdness. I fixed it by returning a pointer to a
Solver, instead of a Solver value. In fact, because a Solver has
internal self-references, it's not safe to copy, so should always be
passed around via pointer.

I wonder if this is a common class of structs that are not copy-safe, or
whether there's some better strategy that doesn't create this problem.